Move local server to separate file
