+++ src/server/index.jsx
@@ -1,26 +1,15 @@
 import express from 'express';
 import compression from 'compression';
-import expressStaticGzip from 'express-static-gzip';
-import path from 'path';
 import ramdaPath from 'ramda/src/path';
 // not part of react-helmet
 import helmet from 'helmet';
 import gnuTP from 'gnu-terry-pratchett';
 import routes from '#app/routes';
 import {
-  articleDataPath,
   articleManifestPath,
   articleSwPath,
-  frontPageDataPath,
   frontPageManifestPath,
   frontPageSwPath,
-  cpsAssetPageDataPath,
-  onDemandRadioDataPath,
-  onDemandTvDataPath,
-  mostReadDataRegexPath,
-  legacyAssetPageDataPath,
-  secondaryColumnDataRegexPath,
-  IdxDataPath,
 } from '../app/routes/utils/regex';
 import nodeLogger from '#lib/logger.node';
 import renderDocument from './Document';
@@ -32,19 +21,17 @@ import {
   MANIFEST_SENDFILE_ERROR,
   SERVER_SIDE_RENDER_REQUEST_RECEIVED,
   SERVER_SIDE_REQUEST_FAILED,
-  LOCAL_SENDFILE_ERROR,
   ROUTING_INFORMATION,
 } from '#lib/logger.const';
 import sendCustomMetric from './utilities/customMetrics';
 import { NON_200_RESPONSE } from './utilities/customMetrics/metrics.const';
+import local from './local';
 
 const fs = require('fs');
 
 const morgan = require('morgan');
 
 const logger = nodeLogger(__filename);
-
-const publicDirectory = 'build/public';
 
 logger.debug(
   `Application outputting logs to directory '${process.env.LOG_DIR}'`,
@@ -56,32 +43,6 @@ class LoggerStream {
     logger.info(message.substring(0, message.lastIndexOf('\n')));
   }
 }
-
-const constructDataFilePath = ({
-  pageType,
-  service,
-  id,
-  variant = '',
-  assetUri,
-}) => {
-  let dataPath;
-
-  switch (pageType) {
-    case 'frontpage':
-    case 'mostRead':
-    case 'secondaryColumn':
-      dataPath = `${variant || 'index'}.json`;
-      break;
-    case 'cpsAssets':
-    case 'legacyAssets':
-      dataPath = `${variant}/${assetUri}.json`;
-      break;
-    default:
-      dataPath = `${id}${variant}.json`;
-  }
-
-  return path.join(process.cwd(), 'data', service, pageType, dataPath);
-};
 
 const server = express();
 
@@ -116,148 +77,17 @@ server
     res.status(200).send(getBuildMetadata());
   });
 
-/*
- * Prod only logging - response time
- */
-if (process.env.SIMORGH_APP_ENV !== 'local') {
-  server.use(logResponseTime);
-}
-
-/*
- * Local env routes - fixture data
- */
-const sendDataFile = (res, dataFilePath, next) => {
-  res.sendFile(dataFilePath, {}, sendErr => {
-    if (sendErr) {
-      logger.error(LOCAL_SENDFILE_ERROR, { error: sendErr });
-      next(sendErr);
-    }
-  });
-};
-
+// Set Up Local Server
 if (process.env.SIMORGH_APP_ENV === 'local') {
-  server
-    .use((req, res, next) => {
-      if (req.url.substr(-1) === '/' && req.url.length > 1)
-        res.redirect(301, req.url.slice(0, -1));
-      else next();
-    })
-    .use(
-      expressStaticGzip(publicDirectory, {
-        enableBrotli: true,
-        orderPreference: ['br'],
-        redirect: false,
-      }),
-    )
-    .get(articleDataPath, async ({ params }, res, next) => {
-      const { service, id, variant } = params;
-
-      const dataFilePath = constructDataFilePath({
-        pageType: 'articles',
-        service,
-        id,
-        variant,
-      });
-
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(frontPageDataPath, async ({ params }, res, next) => {
-      const { service, variant } = params;
-
-      const dataFilePath = constructDataFilePath({
-        pageType: 'frontpage',
-        service,
-        variant,
-      });
-
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(mostReadDataRegexPath, async ({ params }, res, next) => {
-      const { service, variant } = params;
-      const dataFilePath = constructDataFilePath({
-        pageType: 'mostRead',
-        service,
-        variant,
-      });
-
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(onDemandRadioDataPath, async ({ params }, res, next) => {
-      const { service, serviceId, mediaId } = params;
-
-      const dataFilePath = path.join(
-        process.cwd(),
-        'data',
-        service,
-        serviceId,
-        mediaId,
-      );
-
-      sendDataFile(res, `${dataFilePath}.json`, next);
-    })
-    .get(onDemandTvDataPath, async ({ params }, res, next) => {
-      const { service, serviceId, brandEpisode, mediaId } = params;
-
-      const dataFilePath = path.join(
-        process.cwd(),
-        'data',
-        service,
-        serviceId,
-        brandEpisode,
-        mediaId,
-      );
-
-      sendDataFile(res, `${dataFilePath}.json`, next);
-    })
-    .get(cpsAssetPageDataPath, async ({ params }, res, next) => {
-      const { service, assetUri, variant } = params;
-
-      const dataFilePath = constructDataFilePath({
-        pageType: 'cpsAssets',
-        service,
-        assetUri,
-        variant,
-      });
-
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(legacyAssetPageDataPath, async ({ params }, res, next) => {
-      const { service, assetUri, variant } = params;
-
-      const dataFilePath = constructDataFilePath({
-        pageType: 'legacyAssets',
-        service,
-        assetUri,
-        variant,
-      });
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(secondaryColumnDataRegexPath, async ({ params }, res, next) => {
-      const { service, variant } = params;
-      const dataFilePath = constructDataFilePath({
-        pageType: 'secondaryColumn',
-        service,
-        variant,
-      });
-
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get(IdxDataPath, async ({ params }, res, next) => {
-      const { idx } = params;
-      const dataFilePath = path.join(process.cwd(), 'data', idx, 'index.json');
-      sendDataFile(res, dataFilePath, next);
-    })
-    .get('/ckns_policy/*', (req, res) => {
-      // Route to allow the cookie banner to make the cookie oven request
-      // without throwing an error due to not being on a bbc domain.
-      res.sendStatus(200);
-    });
+  local(server);
+} else {
+  // Prod only logging - response time
+  server.use(logResponseTime);
 }
 
 /*
  * Application env routes
  */
-
 server
   .get([articleSwPath, frontPageSwPath], (req, res) => {
     const swPath = `${__dirname}/public/sw.js`;
